pyproject.toml
[project]
name = "lacrei-saude-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Lucas",email = "lucaschaves1370dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
flake8 = "^7.3.0"
black = "^25.1.0"

